split plugin routes generation logic
